chore(Simulation): create environment package
