+++ src/main/scala/ulisse/applications/managers/SimulationManager.scala
@@ -1,8 +1,8 @@
 package ulisse.applications.managers
 
 import ulisse.applications.ports.{SimulationPorts, UtilityPorts}
-import ulisse.entities.simulation.Environments.RailwayEnvironment
 import ulisse.entities.simulation.Simulations.{EngineState, SimulationData}
+import ulisse.entities.simulation.environments.RailwayEnvironment
 import ulisse.utils.Times.{ClockTime, Time}
 
 trait SimulationManager:
