+++ src/main/scala/ulisse/entities/route/Tracks.scala
@@ -2,7 +2,7 @@ package ulisse.entities.route
 
 import cats.data.NonEmptyChain
 import cats.syntax.all.*
-import ulisse.entities.simulation.EnvironmentElements.TrainAgentsContainer
+import ulisse.entities.simulation.environments.EnvironmentElements.TrainAgentsContainer
 import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.utils.CollectionUtils.updateWhen
 import ulisse.utils.Errors.BaseError
