+++ src/main/scala/ulisse/entities/simulation/Simulations.scala
@@ -1,6 +1,6 @@
 package ulisse.entities.simulation
 
-import ulisse.entities.simulation.Environments.RailwayEnvironment
+import ulisse.entities.simulation.environments.RailwayEnvironment
 import ulisse.utils.Times.{ClockTime, Time}
 
 object Simulations:
