+++ src/main/scala/ulisse/entities/timetable/DynamicTimetables.scala
@@ -1,6 +1,6 @@
 package ulisse.entities.timetable
 
-import ulisse.entities.simulation.EnvironmentElements.EnvironmentElement
+import ulisse.entities.simulation.environments.EnvironmentElements.EnvironmentElement
 import ulisse.entities.station.Station
 import ulisse.entities.timetable.Timetables.{StationTime, Timetable}
 import ulisse.utils.Times.{-, ClockTime, InvalidHours, Time, given}
