+++ src/test/scala/ulisse/applications/useCases/SimulationInfoServiceTest.scala
@@ -8,7 +8,7 @@ import ulisse.Runner.runAll
 import ulisse.applications.AppState
 import ulisse.entities.route.RouteEnvironmentElement
 import ulisse.entities.route.Routes.Route
-import ulisse.entities.simulation.Environments.RailwayEnvironment
+import ulisse.entities.simulation.environments.RailwayEnvironment
 import ulisse.entities.station.{Station, StationEnvironmentElement}
 import ulisse.entities.train.TrainAgents.{TrainAgent, TrainAgentInfo}
 import ulisse.entities.train.Trains.Train
